Run each input 10 times (like XCT does with performance tests)
Also format the print out a bit nicer.
